backend: refactor: extract twitterUserObjectScraping.js from worker.js
