+++ backend/src/twitterUserObjectScraping.js
@@ -1,0 +1,133 @@
+const _ = require("lodash")
+const { acquireTwitterApp, getTwitterScreenName } = require("./twitter")
+const { isProduction } = require("./utils")
+const { pgBossQueue, executeInsertOrUpdate } = require("./pg")
+
+/**
+ * @type {string}
+ */
+const TWITTER_USER_OBJECT = "twitterUserObject"
+
+/**
+ * @param {Array<{data: {orgId: string, orgName: string, twitterScreenName: string}}>} jobs
+ *        with data as returned from {@link createTwitterUserObjectScrapingJobData}
+ * @returns {Promise<Array<{orgId: string, orgName: string, twitterUserObject: number}>>}
+ */
+async function scrapeTwitterUserObjects(jobs) {
+  console.log(`Scraping Twitter user objects for: ${JSON.stringify(jobs)}`)
+  if (!isProduction) {
+    return jobs.map(job => ({
+      ...job.data,
+      twitterUserObject: { followers_count: 100 },
+    }))
+  }
+  const screenNames = jobs.map(job => job.data.twitterScreenName).join(",")
+  const app = await acquireTwitterApp()
+  const response = await app.post("users/lookup", {
+    screen_name: screenNames,
+  })
+  const result = _.zipWith(jobs, response, (job, userObject) => ({
+    ...job.data,
+    twitterUserObject: userObject,
+  }))
+  console.log(`Scraped Twitter user objects: ${JSON.stringify(result)}`)
+  return result
+}
+
+/**
+ * @param {Object} org from Airtable
+ * @returns {{orgName: string, twitterUrl: string, orgId: string}}
+ */
+function createTwitterUserObjectScrapingJobData(org) {
+  return {
+    orgId: org.id,
+    orgName: org.fields.Name,
+    twitterScreenName: getTwitterScreenName(org),
+  }
+}
+
+/**
+ * @param {string} jobId
+ * @param {{orgId: string, orgName: string, twitterUserObject: Object}} orgData
+ * @returns {Promise<void>}
+ */
+async function storeTwitterUserObject(jobId, orgData) {
+  try {
+    const now = new Date()
+    await executeInsertOrUpdate(
+      "scraping_results",
+      {
+        org_id: orgData.orgId,
+        request_type: TWITTER_USER_OBJECT,
+        created_at: now,
+      },
+      {
+        updated_at: now,
+        result: orgData.twitterUserObject,
+      }
+    )
+    console.log(
+      `Twitter user object for ${JSON.stringify(
+        orgData
+      )} was successfully stored in the database`
+    )
+    await pgBossQueue.complete(jobId)
+  } catch (err) {
+    console.error(
+      `Error while storing twitter user object for ${JSON.stringify(
+        orgData
+      )} in the database`,
+      err
+    )
+    await pgBossQueue.fail(jobId, err)
+  }
+}
+
+/**
+ * This limit is documented here:
+ * https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup
+ * @type {number}
+ */
+const MAX_ACCOUNTS_PER_TWITTER_USERS_LOOKUP_API_CALL = 100
+
+/**
+ * https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup
+ * documents 300 requests / 15 min window = 1 request / 3 seconds as the limit for app auth. We issue requests more than
+ * three times less frequently than that, 1 request / 10 seconds, just in case, to avoid hitting the limits accidentally
+ * (e. g. when doing integration testing in parallel with production operation using the same app credentials).
+ * @type {number}
+ */
+const DELAY_BETWEEN_TWITTER_USERS_LOOKUP_API_CALLS_MS = 10000
+
+async function twitterUserObjectScrapingLoop() {
+  const jobs = await pgBossQueue.fetch(
+    TWITTER_USER_OBJECT,
+    MAX_ACCOUNTS_PER_TWITTER_USERS_LOOKUP_API_CALL
+  )
+  if (!jobs) {
+    return
+  }
+  let orgsWithTwitterUserObjects
+  try {
+    orgsWithTwitterUserObjects = await scrapeTwitterUserObjects(jobs)
+  } catch (err) {
+    console.error(
+      `Error scraping Twitter user objects for ${JSON.stringify(jobs)}`,
+      err
+    )
+    await pgBossQueue.fail(jobs.map(job => job.id))
+    return
+  }
+  await Promise.all(
+    orgsWithTwitterUserObjects.map((orgData, i) =>
+      storeTwitterUserObject(jobs[i].id, orgData)
+    )
+  )
+}
+
+module.exports = {
+  TWITTER_USER_OBJECT,
+  createTwitterUserObjectScrapingJobData,
+  twitterUserObjectScrapingLoop,
+  DELAY_BETWEEN_TWITTER_USERS_LOOKUP_API_CALLS_MS,
+}
